Done write user api begin write chat_api and his tests
Former-commit-id: 133dba363041e6babcab15db3d201829d16bb6c7

+++ db_api/models.go
@@ -34,7 +34,7 @@ type Chat_User struct {
 	Id   int64`orm:"auto"`
 	User *User`orm:"rel(fk)"`
 	Chat *Chat`orm:"rel(fk)"`
-	Start int`orm:"default(0)"`
+	Start int64`orm:"default(0)"`
 	Delete_last int64`orm:"default(0)"`
 	Delete_points string
 	Ban bool`orm:"default(false)"`
